% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.r
\name{calculate_performance}
\alias{calculate_performance}
\title{Calculate actual performance of methods}
\usage{
calculate_performance(summa, true_val)
}
\arguments{
\item{summa}{A summa object created by the summa() function}

\item{true_val}{A numeric vector of samples corresponding to true labels with 1
representing positive class and -1 representing negative class.}
}
\value{
An object of class summa
}
\description{
This function calculates actual performance of a given set of learners using
a user-defined gold standard. If problem type is binary this calculates
balanced accuracy, if it is of ranks then the performance measure is AUC.
}
\examples{
load(system.file("extdata", "example_rank.Rdata", package = "UnsupervisedEnsemble"))
summa=summa(rank_matrix,"rank")
summa=calculate_performance(summa,actual_labels,,rank)
}
