% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_prediction.R
\name{create_predictions}
\alias{create_predictions}
\title{Create synthetic data}
\usage{
create_predictions(nsamples, nclassifiers, prevelance, type)
}
\arguments{
\item{nsamples}{A number representing the number of samples}

\item{nclassifiers}{A number representing the number of classifiers}

\item{prevelance}{A number representing the percentage of positive samples}

\item{type}{A character either "binary" or "rank"}
}
\value{
A list where the first element is the prediction matrix and the
second element is actual labels
}
\description{
This function creates a synthetic data set as well as the actual labels
}
\examples{
data_binary=create_predictions(100,20,0.4,"binary")
data_rank=create_predictions(100,20,0.7,"rank")
}
