% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor.R
\docType{class}
\name{summa-class}
\alias{summa-class}
\title{An S4 class to represent the summa output.}
\description{
An S4 class to represent the summa output.
}
\section{Slots}{

\describe{
\item{\code{predictions}}{matrix of predictions}

\item{\code{covariance_matrix}}{Square matrix of size learners containing
covariance of learners}

\item{\code{covariance_tensor}}{Cube tensor of size learners containing
the covariance of three learners}

\item{\code{weights}}{Contains weights for each method}

\item{\code{nsamples}}{Number representing number of samples}

\item{\code{nmethods}}{Number representing number of learners}

\item{\code{majority_vote}}{A numeric vector of samples constructed by weighting each
learner equally}

\item{\code{summa}}{A numeric vector of samples constructed by weighting each
learner by their estimated performance}

\item{\code{type}}{A user defined character vector describing the data under analysis}

\item{\code{actual_performance}}{A numeric vector of learnears representing actual
performance of the learners, for binary data this is balanced accuracy
for ranked data this is AUC}

\item{\code{estimated_performance}}{A numeric vector of learners representing
the summa estimated performance of learners}

\item{\code{estimated_prevelance}}{A number corresponding to estimated prevelance}

\item{\code{sampe_rank}}{A numeric vector ranking each sample
by decreased confidence
belonging to positive class using the summa ensemble}
}}

