% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor.R
\name{summa}
\alias{summa}
\title{Construct an unsupervised ensemble}
\usage{
summa(predictions, type)
}
\arguments{
\item{predictions}{A matrix of size samples by methods. If
type="binary", this should consist of binary values. By convention
samples belonging to the negative class should be
denoted by -1 and to the positive denoted by 1. If type="rank" each
entry corresponds to a confidence score
(not necessarily ranked by the user) with samples having higher scores
are more likely to belong to the positive class}

\item{type}{A character specifying the nature of the analysis and can be
either "binary" or "rank".}
}
\value{
An object of class summa
}
\description{
This is the main summa function, which estimates a ranking of a given number
of methods in an unsupervised setting. It also constructs an unsupervised
ensemble of the given methods.
If type is "binary" we implement
the SML method by Parisi et. al. for more detail see [1,2].
}
\examples{
load(system.file("extdata", "example_rank.Rdata", package = "UnsupervisedEnsemble"))
summa=summa(rank_matrix,"rank")
}
\references{
[1] F. Parisi et. al. "Ranking and combining multiple predictors without
labeled data". doi:10.1073, Proceedings of the National Academy of Sciences, 2014.

[2] A. Jaffe et. al. "Estimating the accuracies of multiple classifiers
without labeled data. In AISTATS (Vol. 2,p. 4).
}
