% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_summa.R
\name{summa_plot}
\alias{summa_plot}
\title{Plot predicted versus actual performance}
\usage{
summa_plot(summa)
}
\arguments{
\item{summa}{A summa object created using the calculate_performance() function}
}
\value{
A plot that compares actual v.s. predicted performance as well as
the output of the majority vote and summa ensemble
}
\description{
This is a conveneince function to plot actual vs predicted performances.
Running this function only makes sense if the user knows the true labels and
have already run the calculate_performance() functions using the actual labels.
}
\examples{
data=create_predictions(20,30,0.3,"binary")
summa=summa(data$predictions,"binary")
summa=calculate_performance(summa,data$actual_labels)
summa_plot(summa)
}
